// frontend/src/api/call/models/CallWithTranscript.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Call } from './Call';
import type { TranscriptEntry } from './TranscriptEntry';
export type CallWithTranscript = {
    call: Call;
    transcript: Array<TranscriptEntry>;
};